+++ components/collector/src/source_collectors/sonarqube/suppressed_violations.py
@@ -38,7 +38,7 @@ class SonarQubeSuppressedViolations(SonarQubeViolations):
         branch = self._parameter("branch")
         all_issues_api_url = URL(f"{url}/api/issues/search?projects={component}&branch={branch}")
         resolved_issues_api_url = URL(
-            f"{all_issues_api_url}&resolved=yes&resolutions=WONTFIX,FALSE-POSITIVE&additionalFields=comments"
+            f"{all_issues_api_url}&issuesStatuses=ACCEPTED,FALSE_POSITIVE,IN_SANDBOX&additionalFields=comments"
             f"{self._query_parameter('clean_code_attribute_categories', uppercase=True)}"
             f"{self._query_parameter('impacted_software_qualities', uppercase=True)}"
             f"{self._query_parameter('impact_severities', uppercase=True)}&ps={self.PAGE_SIZE}",
@@ -52,10 +52,9 @@ class SonarQubeSuppressedViolations(SonarQubeViolations):
         return measurement
 
     async def _entity(self, issue) -> Entity:
-        """Extend to add the resolution to the entity."""
+        """Extend to add the issue status to the entity."""
         entity = await super()._entity(issue)
-        resolution = issue.get("resolution", "").lower()
-        entity["resolution"] = {"wontfix": "won't fix"}.get(resolution, resolution)
+        entity["issue_status"] = issue.get("issueStatus", "OPEN").lower().replace("_", " ")
         comments = issue.get("comments", [])
         comments_text = [f"{comment.get('login', 'SonarQube')}: {comment['markdown']}" for comment in comments]
         if comments_text: